requirements.txt
streamlit
easyocr
Pillow
transformers
torch
